Fix warnings for gcc 4.9

+++ arch/arm/mach-msm/bam_dmux.c
@@ -2570,7 +2570,7 @@ static ssize_t show_waketime(struct device *dev,
 	if (!bamDmux_pkt_dev)
 		return 0;
 
-	return snprintf(buf, sizeof(buf), "%u\n", wakelock_timeout);
+	return snprintf(buf, (int)sizeof(buf), "%u\n", wakelock_timeout);
 }
 
 static ssize_t store_waketime(struct device *dev,
